Changed client console logging behavior
The client console will update with logged events from the Python
application.
When the client receives a log message containing the JSON data object,
the output will be truncated to show “Data sent to the server” without
the JSON. This is done to declutter the console.
Additionally, if the console receives a log message after a log message
with the same message body, a counter will increment on the side of the
message to indicate a repeat. This is done to declutter the console.